Выполняет шаги из задания - Учебный проект: разделяй и властвуй
